pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deadlock-tracker"
version = "0.1.0"
description = "Public Deadlock stat tracker website and Discord bot."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.11,<4.0",
    "discord.py>=2.4,<3.0",
    "fastapi>=0.116,<1.0",
    "jinja2>=3.1,<4.0",
    "pillow>=11.2,<13.0",
    "python-dotenv>=1.0,<2.0",
    "python-multipart>=0.0.20,<1.0",
    "uvicorn[standard]>=0.35,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"deadlock_tracker.web" = ["templates/*.html", "static/*.css"]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
